chore: split ExecOffchainConfig definitions
This change splits the ExecOffchainConfig into a versioned
struct (living under v1_0_0) that we parse from JSON and an
internal interface type defined in offramp_reader.go.
It also documents the fields in the config struct.

There are no functional changes.
