+++ core/services/ocr2/plugins/ccip/internal/ccipdata/offramp_reader.go
@@ -19,45 +19,20 @@ const (
 	ManuallyExecute = "manuallyExecute"
 )
 
-// ExecOffchainConfig Do not change the JSON format of this struct without consulting with the RDD people first.
+// ExecOffchainConfig specifies configuration for nodes executing committed messages.
 type ExecOffchainConfig struct {
-	SourceFinalityDepth         uint32
+	// DestOptimisticConfirmations is how many confirmations to wait for the dest chain event before we consider it confirmed (optimistically, need not be finalized).
 	DestOptimisticConfirmations uint32
-	DestFinalityDepth           uint32
-	BatchGasLimit               uint32
-	RelativeBoostPerWaitHour    float64
-	MaxGasPrice                 uint64
-	InflightCacheExpiry         models.Duration
-	RootSnoozeTime              models.Duration
-}
-
-func (c ExecOffchainConfig) Validate() error {
-	if c.SourceFinalityDepth == 0 {
-		return errors.New("must set SourceFinalityDepth")
-	}
-	if c.DestFinalityDepth == 0 {
-		return errors.New("must set DestFinalityDepth")
-	}
-	if c.DestOptimisticConfirmations == 0 {
-		return errors.New("must set DestOptimisticConfirmations")
-	}
-	if c.BatchGasLimit == 0 {
-		return errors.New("must set BatchGasLimit")
-	}
-	if c.RelativeBoostPerWaitHour == 0 {
-		return errors.New("must set RelativeBoostPerWaitHour")
-	}
-	if c.MaxGasPrice == 0 {
-		return errors.New("must set MaxGasPrice")
-	}
-	if c.InflightCacheExpiry.Duration() == 0 {
-		return errors.New("must set InflightCacheExpiry")
-	}
-	if c.RootSnoozeTime.Duration() == 0 {
-		return errors.New("must set RootSnoozeTime")
-	}
-
-	return nil
+	// BatchGasLimit is the maximum sum of user callback gas we permit in one execution report.
+	BatchGasLimit uint32
+	// RelativeBoostPerWaitHour indicates how much to increase (artificially) the fee paid on the source chain per hour of wait time, such that eventually the fee paid is greater than the execution cost and we’ll execute it.
+	// For example: if set to 0.5, that means the fee paid is increased by 50% every hour the message has been waiting.
+	RelativeBoostPerWaitHour float64
+	// InflightCacheExpiry indicates how long we keep a report in the plugin cache before we expire it.
+	// The caching prevents us from issuing another report while one is already in flight.
+	InflightCacheExpiry models.Duration
+	// RootSnoozeTime is the interval at which we check roots for executable messages.
+	RootSnoozeTime models.Duration
 }
 
 type ExecOnchainConfig struct {
